Validating data

+++ Agenda Arterial Sin Arterial/Pacient.swift
@@ -17,38 +17,38 @@ class Pacient {
     var pacientStatus : String
     var birthDate : Date
     var phoneNumber : String
-    var height : Float
-    var weight : Float
-    var cirAbdominal : Float
+    var height : Double
+    var weight : Double
+    var cirAbdominal : Double
     var medDisease : String
     var bloodType : String
     var nextAppointment : Date
     var lastAppointment : Date
     var vinculationCode : String
     var associatedMedic : [String]
-
-    init(id :  String, name : String?, patName : String?, matName : String?, photo : String?, sex : String?, pacientStatus : String?, birthDate : Date, phone : String?,  height : Float?, weight : Float?, cirAbdominal : Float?, medDisease : String?, bloodType : String?,  nextAppointment : Date?, lastAppointment : Date?, vinculationCode : String?, associatedMedic : [String]?){
-    self.id = id
-    self.name = name ?? ""
-    self.patName = patName ?? ""
-    self.matName = matName ?? ""
-    self.photo = photo ?? ""
-    self.sex = sex ?? ""
-    self.pacientStatus = pacientStatus ?? ""
-    self.birthDate = birthDate
-    self.phoneNumber = phone ?? ""
-    self.height = height ?? 0.0
-    self.weight = weight ?? 0.0
-    self.cirAbdominal = cirAbdominal ?? 0.0
-    self.medDisease = medDisease ?? ""
-    self.bloodType = bloodType ?? ""
-    self.nextAppointment = nextAppointment ?? Date()
-    self.lastAppointment = lastAppointment ?? Date()
-    self.vinculationCode = vinculationCode ?? ""
-    self.associatedMedic = associatedMedic ?? []
+    
+    init(id :  String, name : String?, patName : String?, matName : String?, photo : String?, sex : String?, pacientStatus : String?, birthDate : Date, phone : String?,  height : Double?, weight : Double?, cirAbdominal : Double?, medDisease : String?, bloodType : String?,  nextAppointment : Date?, lastAppointment : Date?, vinculationCode : String?, associatedMedic : [String]?){
+        self.id = id
+        self.name = name ?? ""
+        self.patName = patName ?? ""
+        self.matName = matName ?? ""
+        self.photo = photo ?? ""
+        self.sex = sex ?? ""
+        self.pacientStatus = pacientStatus ?? ""
+        self.birthDate = birthDate
+        self.phoneNumber = phone ?? ""
+        self.height = height ?? 0.0
+        self.weight = weight ?? 0.0
+        self.cirAbdominal = cirAbdominal ?? 0.0
+        self.medDisease = medDisease ?? ""
+        self.bloodType = bloodType ?? ""
+        self.nextAppointment = nextAppointment ?? Date()
+        self.lastAppointment = lastAppointment ?? Date()
+        self.vinculationCode = vinculationCode ?? ""
+        self.associatedMedic = associatedMedic ?? []
     }
-
-    func changeInfoPatient(newName : String, newPatName : String, newMatName : String, newPhoto : String, newSex : String, newBirthDate : Date, newPhone : String,  newHeight : Float, newWeight : Float, newCirAbdominal : Float, newMedDisease : String, newBloodType : String,  newNextAppointment : Date) -> (Bool, String){
+    
+    func changeInfoPatient(newName : String, newPatName : String, newMatName : String, newPhoto : String, newSex : String, newBirthDate : Date, newPhone : String,  newHeight : Double, newWeight : Double, newCirAbdominal : Double, newMedDisease : String, newBloodType : String,  newNextAppointment : Date) -> (Bool, String){
         if newName == "" {
             return (false, "Se debe ingresar el nombre")
         }
@@ -73,15 +73,15 @@ class Pacient {
         if newHeight < 0 || newHeight > 999 {
             return (false, "La altura tiene que ser mayor a 0 y menor a 999")
         }
-
+        
         if newWeight  < 0 || newWeight > 999{
             return (false, "El peso tiene que ser mayor a 0 y menor a 999")
         }
-
+        
         if newCirAbdominal  < 0 ||  newCirAbdominal > 999{
             return (false, "La circunferencia abdominal tiene que ser mayor a 0 y menor a 999")
         }
-
+        
         self.name = newName
         self.patName = newPatName
         self.matName = newMatName
@@ -97,7 +97,7 @@ class Pacient {
         self.nextAppointment = newNextAppointment
         return (true, "")
     }
-
+    
 }
